+++ RemotePro/Public/Set-RpRunspaceEventsLegacyVersion.ps1
@@ -310,5 +310,5 @@ function Set-RpRunspaceEventsLegacyVersion {
     # Attach a custom type to RunspaceEvents
     $script:RemotePro.RunspaceEvents.PSTypeNames.Insert(0, 'RemotePro.RunspaceEvents')
 
-    Write-Host "Runspace Events have been successfully added to RemotePro.RunspaceEvents."
+    Write-Verbose "Runspace Events have been successfully added to RemotePro.RunspaceEvents."
 }